And more refactoring!
